add auth

+++ app/Http/Controllers/TasksController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 use App\Task;
 
@@ -15,9 +16,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
-        return view('tasks.index', ['tasks' => $tasks,]);
+        $data = [];
+
+        if(Auth::check()){
+            $user = Auth::user();
+            
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        return view('tasks.index', $data);
     }
 
     /**
@@ -45,10 +57,10 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
         return redirect('/');
     }
@@ -63,6 +75,10 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
         
+        if (Auth::id() !== $task->user_id){
+            return redirect('/');
+        }
+        
         return view('tasks.show', ['task' => $task, ]);
     }
 
@@ -75,6 +91,10 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::findOrFail($id);
+
+        if (Auth::id() !== $task->user_id){
+            return redirect('/');
+        }
         
         return view('tasks.edit', ['task' => $task, ]);
     }
@@ -95,9 +115,11 @@ class TasksController extends Controller
         
         $task = Task::findOrFail($id);
         
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        if(Auth::id() === $task->user_id){
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+        }
         
         return redirect('/');
     }
@@ -112,7 +134,9 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
         
-        $task->delete();
+        if (Auth::id() === $task->user_id){
+            $task->delete();
+        }
         
         return redirect('/');
     }
